pyproject.toml
[tool.poetry]
name = "langchain-sample"
version = "0.0.1"
description = "langchain-sample"
authors = ["rudaks <rudaks94@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<3.13"

langchain = "0.3.2" # 0.1.16
#openai = "==1.21.2"
langchain-openai = "0.2.2" # 0.1.3
langchain-community = "0.3.1"
python-dotenv = "1.0.0"
tiktoken = "*"
chromadb = "*"
black = "*"
langchainhub = "*"
langchain-google-genai = "*"
pillow = "*"
scikit-learn = "*"
faiss-cpu = "*"
lark = "*"
beautifulsoup4 = "*"
langchain-chroma = "*"
sentence-transformers = "*"
langchain-elasticsearch = "*"
numpy = "*"
scipy = "*"
elasticsearch = "*"
pyboxen = "*"
langgraph = "*"
pandas = "*"
langchain-upstage = "*"
langchain-ollama = "*"
langchain-huggingface = "*"
flashrank = "*"
wikipedia = "*"
ipython = "*"
tavily-python = "*"
langchain-experimental = "*"
flagembedding = "*"
peft = "*"
pytesseract = "*"
transformers = "*"
einops = "*"
ninja = "*"
langchain-fireworks = "*"
pdfplumber = "*"
pypdf = "*"
rapidocr-onnxruntime = "*"
pymupdf = "*"
pymupdf4llm = "*"
matplotlib = "*"
openparse = "*"
tesseract = "*"
pysimplegui = "*"
tabula-py = "*"
#camelot-py = {extras=["base"]}
gmft = "*"
llama-index = "*"
llama-parse = "*"
opencv-contrib-python = "*"
tabled = "*"
locust = "*"
uvicorn = {extras = ["standard"], version = "*"}
uvloop = "*"
ocrmypdf = "*"
pdf2image = "*"
jpype1 = "*"
camelot-py = {extras = ["base"], version = "*"}
pypdf2 = "<3.0"
"pdfminer.six" = "*"
ultralyticsplus = "*"
ultralytics = "*"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
